Move util and base classes to org.jboss.as.test.integration.management.base.
