+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/management/api/web/VirtualServerTestCase.java
@@ -38,7 +38,7 @@ import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.arquillian.test.api.ArquillianResource;
-import org.jboss.as.test.integration.management.api.AbstractMgmtTestBase;
+import org.jboss.as.test.integration.management.base.AbstractMgmtTestBase;
 import org.jboss.dmr.ModelNode;
 import org.jboss.shrinkwrap.api.Archive;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
