+++ testsuite/shared/src/main/java/org/jboss/as/test/integration/common/Naming.java
