+++ testsuite/shared/src/main/java/org/jboss/as/test/integration/management/base/AbstractCliTestBase.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.test.integration.management.cli;
+package org.jboss.as.test.integration.management.base;
 
 import java.net.MalformedURLException;
 import java.net.URL;
