+++ testsuite/shared/src/main/java/org/jboss/as/test/integration/management/base/AbstractMgmtTestBase.java
@@ -20,7 +20,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.test.integration.management.api;
+package org.jboss.as.test.integration.management.base;
 
 import java.io.File;
 import java.io.FileInputStream;
