+++ testsuite/shared/src/main/java/org/jboss/as/test/integration/management/util/CLIWrapper.java
