+++ testsuite/shared/src/main/java/org/jboss/as/test/smoke/modular/utils/ShrinkWrapUtils.java
